requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
django-ratelimit>=4.1
django-csp>=3.7
dj-database-url>=2.0
psycopg2-binary>=2.9
gunicorn>=21.2
whitenoise>=6.5
python-dotenv>=1.0
requests>=2.31
pytest>=7.4
pytest-django>=4.5
